add unit tests and reorganize code
